requirements.txt
requests
bs4
logger
codecs
prettytable
jiebas